service2_prime_checker/requirements.txt
nats-py
pydantic
